make the directory more robust
