+++ README.md
@@ -61,15 +61,6 @@ This project uses standard Rust tooling.
     cargo fmt
     ```
 
-### License
-
-This project is dual-licensed under either of:
-
-*   **Apache License, Version 2.0** ([LICENSE-APACHE](./LICENSE-APACHE) or http://www.apache.org/licenses/LICENSE-2.0)
-*   **MIT license** ([LICENSE-MIT](./LICENSE-MIT) or http://opensource.org/licenses/MIT)
-
-at your option.
-
 ### Contribution
 
 We welcome and encourage contributions! Here’s how you can help:
