+++ gui/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "gui"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+wallet_core = { path = "../wallet_core" }
+iced = "0.13.1"
