+++ wallet_core/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "wallet_core"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
